Only return metadata when the requesting user has trustor permissions
Seagull was returning the diagnosis date & birthday of the person
data was shared *with*. This was leaking information which wasn't
explicitly shared.
The only exception is that `fullName` is still returned, as this is used
by Tidepool for Web to render the name of whom you have shared data with
on the "Share" page.
* Fixes https://trello.com/c/oR8PydXP